Fix conceptual bug in Singleton - do not use explicit value but work through strings and create a map from string to value. This map is then transferred to MedFeatures (in attributes) and can be used by OneHot for naming purposes.

+++ Internal/MedProcessTools/MedProcessTools/MedSamples.h
@@ -29,7 +29,7 @@ public:
 
 	/// <summary> Constructor </summary>
 	MedSample() { prediction.clear(); }
-	MedSample(int _id, int _time, float _outcome, float _outcomeTime) { id = _id; time = _time; outcome = _outcome; outcomeTime = _outcomeTime; prediction.clear(); }
+	MedSample(int _id, int _time, int _outcome, int _outcomeTime) { id = _id; time = _time; outcome = _outcome; outcomeTime = _outcomeTime; prediction.clear(); }
 	/// <summary> Destructor </summary>
 	~MedSample() { prediction.clear(); }
